requirements.txt
Django==1.10.4
django-filter==1.0.1
djangorestframework==3.6.2
Markdown==2.6.8
urllib3==1.20